web > status: Fix pluralization of client socket connections
Instead of changing `> 1` to `== 1`, use `yes` variable in consistent
with the rest of the file.

Changes the `connectedClients` type from `int` to `uint8_t` based on the
code review comments in #31.

+++ src/config.h
@@ -58,7 +58,7 @@ struct ConfigSettingsStruct{
   int refreshLogs;
   char hostname[50];
   bool connectedSocket[10];
-  int connectedClients;
+  uint8_t connectedClients;
   unsigned long socketTime;
   int tempOffset;
   bool webAuth;
